Rename some old service.factory files
