+++ src/app/inventory/move-dropped-item.ts
@@ -2,7 +2,7 @@ import { DimStore } from './store-types';
 import { DimItem } from './item-types';
 import { queuedAction } from './action-queue';
 import { reportException } from '../utils/exceptions';
-import { dimItemService } from './dimItemService.factory';
+import { dimItemService } from './item-move-service';
 import { DimError } from '../bungie-api/bungie-service-helper';
 import { t } from 'app/i18next-t';
 import { PlatformErrorCodes } from 'bungie-api-ts/user';
