+++ src/app/inventory/move-item.ts
@@ -3,7 +3,7 @@ import { reportException } from '../utils/exceptions';
 import { queuedAction } from './action-queue';
 import { DimStore } from './store-types';
 import { DimItem } from './item-types';
-import { dimItemService } from './dimItemService.factory';
+import { dimItemService } from './item-move-service';
 import { t } from 'app/i18next-t';
 import { loadingTracker } from '../shell/loading-tracker';
 import { showNotification } from '../notifications/notifications';
